Removed unused references, moved some services into logic project and general cleanup
