+++ src/Carnac.Tests/packages.config
@@ -10,5 +10,4 @@
   <package id="SettingsProviderNet" version="2.1.1" targetFramework="net40" />
   <package id="Shouldly" version="2.4.0" targetFramework="net40" />
   <package id="xunit" version="1.9.2" targetFramework="net40" />
-  <package id="YamlDotNet" version="3.5.1" targetFramework="net40" />
 </packages>
