--- src/Carnac/IShell.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.ObjectModel;
-using Carnac.Logic.Models;
-
-namespace Carnac {
-    public interface IShell {
-        ObservableCollection<Message> Keys { get; }
-        PopupSettings Settings { get; set; }
-    }
-}
