+++ src/KeyStreamCapture/App.xaml.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Configuration;
-using System.Data;
-using System.Linq;
 using System.Windows;
 
 namespace KeyStreamCapture
